+++ CHANGELOG.md
@@ -19,10 +19,15 @@
 - Write environment time to XML in correct format
 - Failing visualization of lanelets, stop lines, traffic signs, and traffic lights with z-coordinate
 - Traffic lights now correctly change size in interactive matplotlib plots (only affected matplotlib>=3.7)
+- Considering state attributes not part of dataclass definition in state to state conversion
+- Enforce InitialState class for initial state property of dynamic obstacle
 
 ### Changed
 - Cleanup lanelet, traffic sign, and traffic light references in function create_from_lanelet_list by default
 - Equality checks of scenario elements no longer emit a warning on inequality (except if the elements are of different types)
+
+### Removed
+- Duplicated initial_state property of dynamic obstacle
 
 ## [2023.2] - 2023-06-26